Description: Tutorial 3: In this tutorial I learned about the beans scope and its lifecycle. I learned the below scopes related to the beans: 1. Singleton(Where single object is shared along the whole app or code) that means whenever we request for create the object then already created object instance is assigned.
2. Prototype(where new instance is created whenever user requested for object or creates the new object).

There are more scopes supported by the Spring but those scopes are supported or used in only web apps.
e.g 1. Request, 2.Session, 3. Global Session etc.

Those scopes will be covered in in-coming sessions.

To add the scope in beans we can use 'scope' attribute in bean tag.

After that I learned about the beans LifeCycle.
i.e. whole flow of beans. See the below:

INITIALIZATION Of Bean --> doing some internal bean processing --> using this bean --> Destroying the Bean.

We can add our own logic or bussiness logic into the INIT of Beans & Destroying of the beans.
e.g. : We can initliaze required DB or file in init method of bean and we can also destroy the DB connection or file in destroy method of the Bean.

To add the init method in bean 'init-method' attribute is used in bean tag.
And to add the destroy method in bean 'destroy-method' attribute is used in bean tag.

Reviewed By:
Prashant Birajdar